clean up public interface
